requirements.txt
langchain
langchain-community
langchain-openai
beautifulsoup4
langchain-chroma
langserve[all]
pydantic
langchain-text-splitters
python-dotenv
pypdf